refactor(store-core): extract plan-mode handlers into handlers/plan.ts (audit P2-3)
Sixth slice of the P2-3 split. Move the plan-mode message handlers
(plan_started / plan_ready / inactivity_warning) and the PlanAllowedPrompt
interface into a new handlers/plan.ts.

Pure move, no logic change: the extracted block is byte-identical to the
removed lines (proven by diff). Imports resolveSessionId + SessionPatch from
./_shared (the enabling slice #5895); PlanAllowedPrompt is local to the
family. The barrel re-exports via `export * from './plan'`, so the public
surface is unchanged.

handlers/index.ts: 5068 -> 4942 lines.
